+++ PhanMemQuanLyChiTieu/app/src/main/java/Adapter/BaoCaoThang.java
@@ -66,7 +66,7 @@ public class BaoCaoThang extends ArrayAdapter<BaoCao> {
         }
         // Object_BaoCaoHienTai item=listData.get(position);
         holder.title.setText(title.get(position));
-        holder.tiencon.setText(String.valueOf(Double.parseDouble(listData.get(position).getTienthu()) - Double.parseDouble(listData.get(position).getTienchi())));
+        holder.tiencon.setText(String.valueOf(Integer.parseInt(listData.get(position).getTienthu()) - Integer.parseInt(listData.get(position).getTienchi())));
         holder.tienthu.setText(listData.get(position).getTienthu());
         holder.tienchi.setText(listData.get(position).getTienchi());
         holder.image.setImageResource(ImageId.get(position));
